PrimePolynom
This gives the solution to the PrimePolynom TopCoder problem at the link https://community.topcoder.com/stat?c=problem_statement&pm=4475&rd=8012

We can check the prime even by Sieve of Eratosthenes but I just applied the O(sqrt(n)) method to make it easier to understand the main crux of the problem.
Basically, we need to find the number m such that when it is used in the given polynomial, it produces a non-prime result. However, all the numbers below m will give prime numbers with the polynomial.